performance optimazation. TODO...

+++ include/micros_swarm_framework/application.h
@@ -86,12 +86,12 @@ namespace micros_swarm_framework{
                 rtp_->setRobotBase(robot_base);
             }
             
-            double getNeighborDistance()
+            float getNeighborDistance()
             {
                 return rtp_->getNeighborDistance();
             }
             
-            void setNeighborDistance(double neighbor_distance)
+            void setNeighborDistance(float neighbor_distance)
             {
                 rtp_->setNeighborDistance(neighbor_distance);
             }
